вынес логику создания новой точки маршрута в модуль trips-screen
